remove tarball and dependent code.
Issues:
Fixes #40

Problem: The tar file in this repo had to be manually generated.

Analysis: Changing a few bits of the patch script removes the
need for this tar ball.

Tests: I tested by listing the contents of a mounted bigip 12.1.x
image during the patch process and verifying that the expected/
necessary files were indeed present in the image.

I then launched the image with tlc, in openstack.  The image booted, licensed and provided a GUI.